joint-map: remove redundant connections[] + add semantic labels
Now that we have a bone-map for describing a hierarchy of bones we can
remove the connections[] state from the joint-map.

This also adds a 'semantic' property to each joint which would
theoretically allow us to maintain alternative joint maps and not rely on
the names of joints to know what actual part of the human body the joint
corresponds to.